pyproject.toml
[tool.poetry]
name = "codefly-sdk"
version = "0.0.7"
description = ""
authors = ["Antoine Toussaint <antoine.toussaint@codefly.ai>"]
packages = [{include = "codefly"}]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.6.3"
PyYAML = "^6.0.1"

[tool.poetry.dev-dependencies]
pytest = "^8.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
